Brenner (#385)
* Add TabithaEngine into the worktree

* Get it compiling, and prune Tabi's actual power

* Remove stat change stuff

* Add base DoR CO

* Enable ability casting via COU

* Add zone logic and zone expansion

* Add charge overrides

* Render CO energy to funds because floating point is suffering

* Add stat effects and display for vets

* Accurately implement the DoR charge bug?

* Actually pass the tests, silly

* Add DoR factions

* Actually add the kill counts tracker

* Add Brenner's files

* Add CO zone display

* Only display my own veterancy

* Enable vet ranks 2 and 3 to function

* Hide zone when COP active, make COP actually work

* Make COUs cost money and have to happen on your property

* Prevent charging during global zone

* Handle the same-size overlay moving around

* Update marks/overlays on COU deploy

* Don't ask me why their default luck is one higher, but it is

* Apply DoR integer math

* Add hooks for ability start/stop so the CO zone updates

* Move zone boosts to the Army level, so tagging is less weird

* Add info about how tags mix with DoR mans

* Don't display the CO zone after the COU's dead

* Don't permit COUing a waited unit

* Totally innocuous, don't worry about it

* Make DoR charge levels more intelligible

* Fix Ave's power cost scaling

* Prevent AIUtils.getAvailableUnitActionsByType() crash

* Text fixes?

* Clarify the ByteBuffer cache invalidation sizing

+++ src/UI/Art/SpriteArtist/SpriteLibrary.java
@@ -624,7 +624,7 @@ public class SpriteLibrary
   }
 
   /** Draw and return an image with the CO's power bar. */
-  public static BufferedImage getCoOverlayPowerBar(Commander co, double maxAP, double currentAP, double pixelsPerPowerUnit)
+  public static BufferedImage getCoOverlayPowerBar(Commander co, int maxAP, double currentAP, double pixelsPerPowerUnit)
   {
     BufferedImage bar = SpriteLibrary.createDefaultBlankSprite((int) (maxAP*pixelsPerPowerUnit), 5);
     Graphics barGfx = bar.getGraphics();
