+++ src/Units/AWBWUnits.java
@@ -140,7 +140,7 @@ public class AWBWUnits extends UnitModelScheme
     }
 
     @Override
-    public double getDamageRedirect(WeaponModel wm)
+    public int getDamageRedirect(WeaponModel wm)
     {
       return wm.getDamage(this);
     }
