+++ src/Units/ITargetable.java
@@ -3,5 +3,5 @@ package Units;
 public interface ITargetable
 {
   /** Calls the appropriate type-specific override of getDamage() on the input WeaponModel */
-  public abstract double getDamageRedirect(WeaponModel wm);
+  public abstract int getDamageRedirect(WeaponModel wm);
 }
