+++ src/Units/KaijuWarsUnits.java
@@ -281,7 +281,7 @@ public class KaijuWarsUnits extends UnitModelScheme
     }
 
     @Override
-    public double getDamageRedirect(WeaponModel wm)
+    public int getDamageRedirect(WeaponModel wm)
     {
       return wm.getDamage(this);
     }
